Adjust CODE_128 to handle odd number of digits
This is a change so that handling for CODE_128 auto takes into account an odd number of digits after an initial string. This then works for international DPD shipping integration "NOTE: If the postcode is all numeric, use subset B for the first 2 digits of the data
string, e.g. %0, followed by subset C for the rest of the barcode."

Example %008099915501071048275101276
Previously this would output
 * SUBSET B = %, length 1
 * SUBSET C = 00809991550107104827510127, length 26
 * SUBSET B = 6, length 1
Now expected outcome is
 * SUBSET B = %0, length 2
 * SUBSET C = 08099915501071048275101276, length 26